# public/blgpt-openapi.yaml
openapi: 3.1.0
info:
  title: BL-GPT Binance Analysis API
  version: 1.0.0
servers:
  - url: https://bl-gpt-server.onrender.com
paths:
  /price/{symbol}:
    get:
      summary: Coin fiyatını getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fiyat bilgisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  price:
                    type: string

  /rsi/{symbol}:
    get:
      summary: RSI verisi getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: RSI değeri
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  rsi:
                    type: number

  /macd/{symbol}:
    get:
      summary: MACD sinyali getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MACD verisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  macd:
                    type: string

  /ema/{symbol}:
    get:
      summary: EMA 9 ve 21 değerlerini getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: EMA verileri
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  ema9:
                    type: number
                  ema21:
                    type: number

  /bollinger/{symbol}:
    get:
      summary: Bollinger bandı verilerini getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bollinger verisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  upper:
                    type: number
                  middle:
                    type: number
                  lower:
                    type: number
                  breakout:
                    type: string

  /atr/{symbol}:
    get:
      summary: ATR verisini getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ATR değeri
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  atr:
                    type: number

  /orderbook/{symbol}:
    get:
      summary: Order book verisini getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alış-satış emirleri
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  bids:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
                  asks:
                    type: array
                    items:
                      type: array
                      items:
                        type: string

  /funding/{symbol}:
    get:
      summary: Funding rate getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Funding rate değeri
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  fundingRate:
                    type: string

  /volume/{symbol}:
    get:
      summary: Hacim verisi getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hacim bilgisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  volume:
                    type: string

  /volatility/{symbol}:
    get:
      summary: Volatilite seviyesini getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Volatilite bilgisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  volatility:
                    type: string

  /risk/{symbol}:
    get:
      summary: Risk skoru getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risk skoru
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  riskScore:
                    type: number

  /fundamentals/{symbol}:
    get:
      summary: Temel analiz özeti getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Temel analiz bilgisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  summary:
                    type: string

  /onchain/{symbol}:
    get:
      summary: Zincir üstü verileri getir
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: On-chain veriler
          content:
            application/json:
              schema:
                type: object
                properties:
                  coin:
                    type: string
                  activeAddresses:
                    type: integer
                  whaleHoldingPercent:
                    type: number
  /analyze/top:
    get:
      summary: En güçlü long sinyali veren coinleri getir
      description: RSI, MACD, EMA, Bollinger ve Temel analiz skorlarına göre en uygun 3 coini listeler
      responses:
        '200':
          description: En yüksek puanlı ilk 3 coin
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    coin:
                      type: string
                      example: "SOLUSDT"
                    score:
                      type: integer
                      example: 5

  /analyze/full:   # 👈 Doğru hizalanmış!
    get:
      summary: Gelişmiş analiz – 12 göstergeye göre en uygun coinleri getir
      responses:
        '200':
          description: En güçlü sinyale sahip 5 coin
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    coin:
                      type: string
                      example: SOLUSDT
                    direction:
                      type: string
                      enum: [long, short]
                    score:
                      type: integer
                      example: 7
                    reasons:
                      type: array
                      items:
                        type: string
  /news/{symbol}:
    get:
      summary: Binance Blog'dan belirli bir coin ile ilgili haberleri getir
      description: Coin ismine göre Binance resmi blog RSS'inden filtrelenmiş en güncel haberleri listeler.
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
          example: SOL
      responses:
        '200':
          description: Coin ile ilgili en güncel haberler
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                      example: "Solana: Yeni Gelişmeler"
                    summary:
                      type: string
                      example: "Solana ekibi tarafından açıklanan yeni staking sistemi..."
                    link:
                      type: string
                      example: "https://www.binance.com/en/blog/post1234"
                    date:
                      type: string
                      example: "2025-07-25T08:30:00Z"